requirements.txt
click
numpy
pytest
pandas
matplotlib
biopython
scikit_learn
scipy
